backend/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
psycopg2-binary
